+++ src/main/java/com/kr1s1s/subtlyd/SubtlyDungeons.java
@@ -1,20 +1,14 @@
 package com.kr1s1s.subtlyd;
 
 import com.kr1s1s.subtlyd.client.util.GroundShake;
-import com.kr1s1s.subtlyd.world.entity.EntitySD;
-import com.kr1s1s.subtlyd.world.entity.TentEntity;
 import com.kr1s1s.subtlyd.world.item.ItemsSD;
 import net.fabricmc.api.ModInitializer;
 
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
-import net.fabricmc.fabric.api.object.builder.v1.entity.FabricDefaultAttributeRegistry;
-import net.minecraft.core.Registry;
-import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.network.syncher.EntityDataAccessor;
 import net.minecraft.network.syncher.EntityDataSerializers;
 import net.minecraft.network.syncher.SynchedEntityData;
 import net.minecraft.world.entity.monster.Zombie;
-import net.minecraft.world.item.Item;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -38,6 +32,5 @@ public class SubtlyDungeons implements ModInitializer {
         });
 
         ItemsSD.registerItems();
-        //EntitySD.registerEntities();
     }
 }