+++ src/main/java/com/kr1s1s/subtlyd/client/model/TentModel.java
@@ -1,29 +1,17 @@
-package com.kr1s1s.subtlyd.client.entity.render.model;
+package com.kr1s1s.subtlyd.client.model;
 
-import com.kr1s1s.subtlyd.SubtlyDungeons;
-import com.kr1s1s.subtlyd.client.entity.render.layers.TentRenderState;
+import com.kr1s1s.subtlyd.client.renderer.state.TentRenderState;
 import net.minecraft.client.model.EntityModel;
-import net.minecraft.client.model.geom.ModelLayerLocation;
 import net.minecraft.client.model.geom.ModelPart;
 import net.minecraft.client.model.geom.PartPose;
 import net.minecraft.client.model.geom.builders.CubeListBuilder;
 import net.minecraft.client.model.geom.builders.LayerDefinition;
 import net.minecraft.client.model.geom.builders.MeshDefinition;
 import net.minecraft.client.model.geom.builders.PartDefinition;
-import net.minecraft.resources.ResourceLocation;
 
 public class TentModel extends EntityModel<TentRenderState> {
-    public static final ModelLayerLocation LAYER_LOCATION = new ModelLayerLocation(
-            ResourceLocation.fromNamespaceAndPath(SubtlyDungeons.MOD_ID, "white_tent"), "main"
-    );
-
-    private final ModelPart root;
-    private final ModelPart TentModel;
-
     public TentModel(ModelPart root) {
         super(root);
-        this.root = root;
-        this.TentModel = root.getChild("TentModel");
     }
 
     public static LayerDefinition createBodyLayer() {