+++ src/main/java/com/kr1s1s/subtlyd/client/model/mob/zombie/DrownedModelSD.java
@@ -1,10 +1,8 @@
-package com.kr1s1s.subtlyd.client.entity.render.model.mob.zombie;
+package com.kr1s1s.subtlyd.client.model.mob.zombie;
 
-import com.kr1s1s.subtlyd.client.entity.render.layers.zombie.ZombieRenderStateSD;
-import net.minecraft.client.model.HumanoidModel;
+import com.kr1s1s.subtlyd.client.renderer.state.ZombieRenderStateSD;
 import net.minecraft.client.model.ZombieModel;
 import net.minecraft.client.model.geom.ModelPart;
-import net.minecraft.client.model.geom.PartPose;
 import net.minecraft.client.model.geom.builders.*;
 import net.minecraft.util.Mth;
 
@@ -13,27 +11,15 @@ public class DrownedModelSD extends ZombieModel<ZombieRenderStateSD> {
         super(modelPart);
     }
 
-    public static LayerDefinition createBodyLayer(CubeDeformation cubeDeformation) {
-        MeshDefinition meshDefinition = HumanoidModel.createMesh(cubeDeformation, 0.0F);
-        PartDefinition partDefinition = meshDefinition.getRoot();
-        partDefinition.addOrReplaceChild(
-                "left_arm", CubeListBuilder.create().texOffs(32, 48).addBox(-1.0F, -2.0F, -2.0F, 4.0F, 12.0F, 4.0F, cubeDeformation), PartPose.offset(5.0F, 2.0F, 0.0F)
-        );
-        partDefinition.addOrReplaceChild(
-                "left_leg", CubeListBuilder.create().texOffs(16, 48).addBox(-2.0F, 0.0F, -2.0F, 4.0F, 12.0F, 4.0F, cubeDeformation), PartPose.offset(1.9F, 12.0F, 0.0F)
-        );
-        return LayerDefinition.create(meshDefinition, 64, 64);
-    }
-
     @Override
     public void setupAnim(ZombieRenderStateSD zombieRenderState) {
         super.setupAnim(zombieRenderState);
-        if (zombieRenderState.leftArmPose == HumanoidModel.ArmPose.THROW_SPEAR) {
+        if (zombieRenderState.leftArmPose == ArmPose.THROW_SPEAR) {
             this.leftArm.xRot = this.leftArm.xRot * 0.5F - (float) Math.PI;
             this.leftArm.yRot = 0.0F;
         }
 
-        if (zombieRenderState.rightArmPose == HumanoidModel.ArmPose.THROW_SPEAR) {
+        if (zombieRenderState.rightArmPose == ArmPose.THROW_SPEAR) {
             this.rightArm.xRot = this.rightArm.xRot * 0.5F - (float) Math.PI;
             this.rightArm.yRot = 0.0F;
         }