+++ src/main/java/com/kr1s1s/subtlyd/client/renderer/state/TentRenderState.java
@@ -1,7 +1,6 @@
-package com.kr1s1s.subtlyd.client.entity.render.layers;
+package com.kr1s1s.subtlyd.client.renderer.state;
 
 import net.minecraft.client.renderer.entity.state.EntityRenderState;
-import net.minecraft.client.renderer.entity.state.LivingEntityRenderState;
 import net.minecraft.util.Mth;
 import net.minecraft.world.item.DyeColor;
 
@@ -10,7 +9,6 @@ public class TentRenderState extends EntityRenderState {
     public boolean hasRedOverlay;
     public float yRot;
     public float xRot;
-    public DyeColor color;
 
     public float getXRot(float f) {
         return f == 1.0F ? this.xRot : Mth.lerp(f, this.xRot, this.xRot);